Move Piano component
